gfxlib2: Adjust fb_hScreenInfo() parameters for 64bit
Since 5f2f62ea screeninfo() uses INTEGER (ssize_t) parameters; so one 64bit
helper functions used by fb_GfxScreenInfo(), such as fb_hScreenInfo(), must
be adjusted to match that (to prevent pointer type mismatches)

+++ src/gfxlib2/dos/gfx_dos.c
@@ -662,7 +662,7 @@ static void fb_dos_restore_video_mode(void)
 	fb_dos.old_rows = fb_dos.old_cols = 0;
 }
 
-void fb_hScreenInfo(int *width, int *height, int *depth, int *refresh)
+void fb_hScreenInfo(ssize_t *width, ssize_t *height, ssize_t *depth, ssize_t *refresh)
 {
 	*width = fb_dos.w;
 	*height = fb_dos.h;
